+++ agendamlg-war/src/java/service/Agendamlg.java
@@ -137,7 +137,7 @@ public class Agendamlg {
     }
     
     @WebMethod(operationName = "buscarEventoCategorias")
-    public List<app.entity.Evento> buscarEventoCategorias(@WebParam(name = "categorias")List<app.entity.Categoria> categorias, @WebParam(name = "usuario")app.entity.Usuario usuario, @WebParam(name = "filtrarCategorias")boolean filtroCercania, @WebParam(name = "coordenadaX")int x, @WebParam(name = "coordenadaY")int y, @WebParam(name = "radio")int radio){
+    public List<app.entity.Evento> buscarEventoCategorias(@WebParam(name = "categorias")List<app.entity.Categoria> categorias, @WebParam(name = "usuario")app.entity.Usuario usuario, @WebParam(name = "filtrarCategorias")boolean filtroCercania, @WebParam(name = "coordenadaX")double x, @WebParam(name = "coordenadaY")double y, @WebParam(name = "radio")double radio){
         if(usuario != null) usuario = usuarioFacade.find(usuario.getId());
         return eventoFacade.buscarEventoCategorias(categorias, usuario, filtroCercania, x, y, radio);
     }
